Fixes #28530: Combine post install hooks into a single one
This moves to a single post install hook to handle all scenarios.
This introduces the message helpers to Foreman hooks and starts
to provide the hook extensions Katello uses.
