--- katello/checks
@@ -1,1 +1,0 @@
-../checks
