--- katello/hooks/post/10-post_install.rb
@@ -1,39 +1,0 @@
-def success_file
-  File.join(File.dirname(Kafo::KafoConfigure.config_file), '.installed')
-end
-
-def new_install?
-  !File.exist?(success_file)
-end
-
-def proxy?
-  system("rpm -q foreman-proxy-content > /dev/null") && !system("rpm -q katello > /dev/null")
-end
-
-if [0, 2].include?(@kafo.exit_code)
-  if !app_value(:upgrade)
-    if module_enabled?('katello')
-      Kafo::MessageHelpers.server_success_message(@kafo)
-      Kafo::MessageHelpers.new_install_message(@kafo) if new_install?
-      Kafo::MessageHelpers.certs_generate_command_message
-    elsif module_enabled?('katello_devel')
-      Kafo::MessageHelpers.dev_server_success_message(@kafo)
-      Kafo::MessageHelpers.dev_new_install_message(@kafo) if new_install?
-    end
-
-    Kafo::MessageHelpers.proxy_success_message(@kafo) if proxy?
-  end
-
-  File.open(success_file, 'w') {} unless app_value(:noop) # Used to indicate that we had a successful install
-  exit_code = 0
-
-else
-
-  say "  <%= color('Something went wrong!', :bad) %> Check the log for ERROR-level output"
-  exit_code = @kafo.exit_code
-
-end
-
-# This is always useful, success or fail
-log = @kafo.config.app[:log_dir] + '/' + @kafo.config.app[:log_name]
-say "  The full log is at <%= color('#{log}', :info) %>"
